chore: move internal to another file
